refactor: ensure app uses TS baseUrl to improve imports dev experience
