+++ jest.config.js
@@ -7,4 +7,8 @@ module.exports = {
   transform: {
     ".+\\.ts$": "ts-jest",
   },
+  moduleNameMapper: {
+    "@/tests/(.*)": "<rootDir>/tests/$1",
+    "@/(.*)": "<rootDir>/src/$1",
+  },
 };
