+++ src/data/services/patient-data-min-max-calculator-service.ts
@@ -1,6 +1,6 @@
-import type { Patient } from "../../domain/models";
-import type { MinMaxPatientValues } from "../../domain/protocols";
-import type { IPatientDataMinMaxCalculatorService } from "../../domain/services";
+import type { Patient } from "@/domain/models";
+import type { MinMaxPatientValues } from "@/domain/protocols";
+import type { IPatientDataMinMaxCalculatorService } from "@/domain/services";
 
 export class PatientDataMinMaxCalculatorService
   implements IPatientDataMinMaxCalculatorService
