+++ src/data/services/patient-score-calculator-service.ts
@@ -1,9 +1,9 @@
 import type {
   IPatientDataNormalizerService,
   IPatientScoreCalculatorService,
-} from "../../domain/services";
-import type { Patient } from "../../domain/models";
-import type { MinMaxPatientValues } from "../../domain/protocols";
+} from "@/domain/services";
+import type { Patient } from "@/domain/models";
+import type { MinMaxPatientValues } from "@/domain/protocols";
 
 export class PatientScoreCalculatorService
   implements IPatientScoreCalculatorService
