+++ src/main/adapters/express-route-adapter.ts
@@ -1,6 +1,6 @@
 import type { Request, Response } from "express";
 
-import { Controller } from "../../presentation/protocols";
+import { Controller } from "@/presentation/protocols";
 
 export const adaptRoute = (controller: Controller) => {
   return async (req: Request, res: Response) => {
