+++ src/main/config/middlewares.ts
@@ -1,8 +1,6 @@
 import type { Express } from "express";
 
-import { cors } from "../middlewares/cors";
-import { bodyParser } from "../middlewares/body-parser";
-import { contentType } from "../middlewares/content-type";
+import { cors, bodyParser, contentType } from "@/main/middlewares";
 
 export default (app: Express) => {
   app.use(cors);
