+++ src/main/config/routes.ts
@@ -1,6 +1,6 @@
 import { Express, Router } from "express";
 
-import PatientsRoutes from "../routes/patients-routes";
+import PatientsRoutes from "@/main/routes/patients-routes";
 
 export default (app: Express) => {
   const router = Router();
