+++ src/main/config/swagger.ts
@@ -1,7 +1,7 @@
 import { serve, setup } from "swagger-ui-express";
 import { Express } from "express";
 
-import swaggerConfig from "../docs";
+import swaggerConfig from "@/main/docs";
 
 export default (app: Express): void => {
   app.use("/api-docs", serve);
