+++ tests/main/middlewares/content-type.test.ts
@@ -1,6 +1,6 @@
 import request from "supertest";
 
-import { app } from "../../../src/main/config";
+import { app } from "@/main/config";
 
 describe("contentType middleware", () => {
   it("Should return default content type as json", async () => {
