+++ tests/main/routes/patients-routes.test.ts
@@ -1,6 +1,6 @@
 import request from "supertest";
 
-import { app } from "../../../src/main/config";
+import { app } from "@/main/config";
 
 describe("Patients Routes", () => {
   describe("GET /patients-recommended", () => {
