+++ tests/main/services/mocks/mock-patient-data-normalizer-service.ts
@@ -1,6 +1,6 @@
-import type { Patient } from "../../../domain/models";
-import type { MinMaxPatientValues } from "../../../domain/protocols";
-import type { IPatientDataNormalizerService } from "../../../domain/services";
+import type { Patient } from "@/domain/models";
+import type { MinMaxPatientValues } from "@/domain/protocols";
+import type { IPatientDataNormalizerService } from "@/domain/services";
 
 export class PatientDataNormalizerSpy implements IPatientDataNormalizerService {
   normalizeFieldInput?: any;
